translate/requirements.txt
googletrans==4.0.0rc1
gTTS
djangorestframewok
django-cors-headers
